Unify fsb/wwise/ogl Vorbis decoders into a single decoder with subtypes
